Add ability to attach a spotlight to an agent
Calling Model.setSpotlight(agent) will add a "spotlight"
to an agent, similar to the NetLogo "follow" function.
This allows a user to easily follow a single agent.

The spotlight is removed when Model.removeSpotlight()
is called, or when the agent dies.

There is also a shortcut for Model.setSpotlight(breed),
which will pick a random member of a breed.